setup for problem description
